AYS-28 | `InvalidTokenDeletionSchedulerSystemTest` Has Been Refactored
