+++ src/test/resources/application.yml
@@ -20,6 +20,6 @@ spring:
 
 ays:
   scheduler:
-    UnusedInvalidTokenDeletionScheduler:
-      cron: '0/1 * * ? * *'
-      enabled: true
+    invalid-tokens-deletion:
+      cron: 0/1 * * ? * *
+      enable: true
